Fix table of contents
The table of contents in the README was using a h2 section headline, which can be confused in some auto table of content creators as a section, which messes up the contents. Here, I change it to just a bold text so that it doesn't get confused in the future.